More than one instance found is now an error
* Move errors functions/vars into capoerrors, so they can easily be
  re-used.
* Now we return an error if multiple instances with the same name were
  found (and not just a warning). There is no safe course of action here.
